+++ HISTORY.rst
@@ -26,3 +26,4 @@ History
 * Generate cid outside of the middleware when ``GENERATE_CID`` is
   enabled, so that it's available even if the middleware is not used
 * Fix support of Django 1.11 in database backends.
+* Add PostGIS database backend.